requirements.txt
# Core web scraping dependencies
Crawl4AI==0.4.247
pydantic>=2.4.2
python-dotenv>=1.0.1

# Content extraction and processing
trafilatura>=1.6.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Streamlit frontend dependencies
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
